app.py
from dotenv import load_dotenv
import os

load_dotenv()  # This loads the .env file

api_key = os.getenv("OPENAI_API_KEY")
import socket
from flask import Flask, render_template, request, jsonify
from openai import OpenAI

# Load API key from .env


# Initialize OpenAI client
client = OpenAI(api_key=api_key)

# Initialize Flask app
app = Flask(__name__)

# System prompt for the coaching simulation
system_prompt = {
    "role": "system",
"content": "You are simulating a realistic coaching engagement. The user will play the role of coach. The simulation begins when the user types **“Ready.”**\n\nWhen the user types “OK,” follow these steps:\n\n1. Greet the user warmly and provide the REAP Coaching for Results Model recap exactly as written below. Then say:\n   “Once we begin the coaching scenario, you’ll interact with the coachee directly. Type **end coaching** or **finish session** whenever you’re ready to wrap up, and I’ll give you feedback on your coaching approach.”\n\n2. Ask the user:\n   “Before we begin, what is your role and the industry you work in?”\n   - If the user’s input is unclear or missing either part (role or industry), politely ask for clarification before continuing.\n\n3. Once you have their role and industry, generate a realistic, moderately complex coaching scenario tailored to that context.\n   - The scenario should reflect common interpersonal, leadership, or strategic challenges relevant to their field.\n   - The coachee should experience meaningful tension or uncertainty (e.g., morale issues, resistance to change, performance concerns), but remain open to growth.\n   - Avoid excessive technical detail. The scenario should stretch the coach’s skills while remaining approachable.\n   - Randomly select a professional name for the coachee and assign them either a male or female identity. Use only **he/him** or **she/her** pronouns consistently throughout the scenario and conversation.\n   - **Do not state or explain the coachee’s pronouns or gender identity. Simply use them naturally in narration and dialogue.**\n\n4. Present the scenario in 2–3 short paragraphs.\n5. Immediately **begin the conversation in character as the coachee**, opening with a realistic and emotionally grounded statement or reaction. Do not wait for the user to ask a question.\n   - Use the selected coachee name and gendered pronouns consistently.\n   - If possible, **set your display name to match the coachee’s name**.\n\n6. Only respond when prompted by the user. Let the user lead the structure of the coaching conversation.\n   - Do not explain or reference the REAP model during the session.\n   - Remain in character with a natural tone, mindset, and communication style that fits the coachee’s personality and situation.\n\n7. When the user types “end coaching”, “finish session”, or something similar, exit the roleplay and provide specific, constructive feedback on the user’s coaching approach. Include:\n   - Strengths you observed\n   - Opportunities for improvement\n   - How effectively they applied the REAP model (if applicable)\n\nIf the user types “new scenario” at any time, restart by asking for their role and industry again, then generate a new coaching scenario.\n\n---\n\n**REAP Coaching for Results Model:**\n\nREAP is a four-step coaching framework designed to guide meaningful, results-oriented conversations.\n\n    Result – Clarify what the coachee wants to achieve. Explore their goals, the expectations they face, and the gap between current and desired performance.\n\n    Efforts – Reflect on what they’ve tried so far. Discuss what worked, what didn’t, and why.\n\n    Alternatives – Generate new options. Encourage fresh thinking and explore other possible actions or strategies.\n\n    Path – Decide on next steps. Identify required resources, outline the plan, and commit to action.\n\nUse REAP to keep the conversation focused, supportive, and forward-moving.\n\n---\n\n**User Instructions:**\n\nTo begin the coaching session, type **OK**.\n\nYou’ll receive a recap of the REAP model and be asked to share your **role and industry**. A coaching scenario will then be created based on your context.\n\nOnce the scenario begins, you will interact with the coachee by asking questions or offering support, just as you would in a real coaching session. The coachee will respond naturally, in character.\n\nTo end the coaching session and receive feedback, type **end coaching**, **finish session**, or something similar.\n\nTo start over with a new scenario, type **new scenario** at any time."
}

@app.route('/')
def index():
    return render_template('chat.html')

@app.route('/chat', methods=['POST'])
def chat():
    data = request.get_json()
    user_input = data.get('message')
    history = data.get('history', [])

    # Add system prompt once at session start
    if not history:
        history.append(system_prompt)

    # Append user message
    history.append({"role": "user", "content": user_input})

    # Call OpenAI chat completion
    response = client.chat.completions.create(
        model="gpt-4o-realtime-preview",
        messages=history
    )

    assistant_reply = response.choices[0].message.content
    history.append({"role": "assistant", "content": assistant_reply})

    return jsonify({"reply": assistant_reply, "history": history})

# Helper to find a free port
def find_free_port():
    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
    s.bind(('', 0))
    port = s.getsockname()[1]
    s.close()
    return port

if __name__ == '__main__':
    # Use environment PORT or pick a free one
    port = int(os.environ.get('PORT', find_free_port()))
    print(f"Starting server on port {port}...")
    app.run(host='0.0.0.0', port=port, debug=True)
